# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 伴AI API
  version: 1.27.0
  description: M1-M4 platform plus the M6 reliability control plane, governed Agent Studio, unified adjustable usage ledger, multi-instance runtime-configuration convergence, persisted cost and quality analytics with time trends, predictive budget burn and guarded model-switch recommendations, per-budget proactive forecast policies with sample gating, audited human recommendation feedback, revision-bound application and rollback-execution tracking, configurable before/after observation windows, downloadable periodic budget reviews, read-only rollback recommendations and post-rollback verification with acknowledged/closed human disposition, audited budget guardrails, automatic budget incidents with escalation plus email or management-console recovery notifications, baseline anomaly detection, Loki-backed redacted structured system logs with PostgreSQL fallback and Run/Trace correlation, threshold alert evaluation, downloadable incident evidence, durable W3C/OpenTelemetry Trace Context propagation across HTTP, Outbox, Kafka, Go/Python Agent workers and Langfuse, optional fail-open Langfuse LLM/Agent traces with generations, node observations and scores, operator DLQ replay, and compensation records. Every HTTP response includes canonical traceparent and X-Trace-ID correlation headers; clients may send traceparent/tracestate, while X-Trace-ID remains a normalized compatibility input.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/ops/performance/versions:
    get:
      operationId: listOperatorVersionPerformance
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Aggregates privacy-safe success, quality, latency, token, and cost metrics by immutable Agent or model-profile version.
      parameters:
        - { name: dimension, in: query, required: false, schema: { type: string, enum: [agent_version, model_profile], default: agent_version } }
        - { name: module, in: query, required: false, schema: { type: string, enum: [companion, life, work] } }
        - { name: range, in: query, required: false, schema: { type: string, enum: [24h, 7d, 30d, 90d], default: 7d } }
        - { name: from, in: query, required: false, description: Must be supplied together with to; overrides range., schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, description: Must be supplied together with from; overrides range., schema: { type: string, format: date-time } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        "200":
          description: Version metrics ordered by completed sample count and recency.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VersionPerformancePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/performance/models:
    get:
      operationId: listOperatorModelPerformance
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Aggregates the provider and model actually returned for persisted Agent model calls, including failures and cost.
      parameters:
        - { name: module, in: query, required: false, schema: { type: string, enum: [companion, life, work] } }
        - { name: range, in: query, required: false, schema: { type: string, enum: [24h, 7d, 30d, 90d], default: 7d } }
        - { name: from, in: query, required: false, description: Must be supplied together with to; overrides range., schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, description: Must be supplied together with from; overrides range., schema: { type: string, format: date-time } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        "200":
          description: Actual provider/model usage ordered by total cost and call volume.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelPerformancePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/performance/anomalies:
    get:
      operationId: getOperatorPerformanceAnomalies
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Compares a recent completed-run window with the immediately preceding baseline and suppresses decisions when sample counts are too small.
      parameters:
        - { name: module, in: query, required: false, schema: { type: string, enum: [companion, life, work] } }
        - { name: window, in: query, required: false, schema: { type: string, enum: [1h, 6h, 24h], default: 1h } }
        - { name: baseline, in: query, required: false, schema: { type: string, enum: [24h, 7d, 30d], default: 24h } }
      responses:
        "200":
          description: Current and baseline metrics with any statistically eligible threshold violations.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PerformanceAnalysis" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/performance/trend:
    get:
      operationId: listOperatorPerformanceTrend
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Returns privacy-safe cost, success, call-volume, and latency buckets. Ranges up to seven days use hourly buckets; longer ranges use daily buckets.
      parameters:
        - { name: module, in: query, required: false, schema: { type: string, enum: [companion, life, work] } }
        - { name: range, in: query, required: false, schema: { type: string, enum: [24h, 7d, 30d, 90d], default: 7d } }
        - { name: from, in: query, required: false, description: Must be supplied together with to; overrides range., schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, description: Must be supplied together with from; overrides range., schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Chronological trend buckets including empty buckets.
          content: { application/json: { schema: { $ref: "#/components/schemas/PerformanceTrendPage" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/performance/forecast:
    get:
      operationId: getOperatorPerformanceForecast
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Projects each matching budget to the end of its current period using actual burn rate over the selected history window. It also proposes lower-cost observed models only when sample, error-rate, and P95-latency guardrails pass. Recommendations are read-only and require validation before adoption.
      parameters:
        - { name: module, in: query, required: false, schema: { type: string, enum: [companion, life, work] } }
        - { name: range, in: query, required: false, schema: { type: string, enum: [24h, 7d, 30d, 90d], default: 7d } }
        - { name: from, in: query, required: false, description: Must be supplied together with to; overrides range., schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, description: Must be supplied together with from; overrides range., schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Read-only budget and model optimization forecast.
          content: { application/json: { schema: { $ref: "#/components/schemas/PerformanceForecast" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/performance/budgets:
    get:
      operationId: listOperatorPerformanceBudgets
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Evaluates persisted daily and monthly cost guardrails against actual Agent model-call costs and returns each budget's proactive forecast policy.
      responses:
        "200":
          description: Budget definitions with current-period utilization and status.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [budgets]
                properties:
                  budgets: { type: array, items: { $ref: "#/components/schemas/PerformanceBudgetStatus" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createOperatorPerformanceBudget
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Creates one audited budget per module scope and period. Requires the admin role.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PerformanceBudgetMutation" } } }
      responses:
        "201":
          description: Budget created.
          content:
            application/json:
              schema:
                type: object
                required: [budget]
                properties: { budget: { $ref: "#/components/schemas/PerformanceBudget" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/performance/budgets/forecast-history:
    get:
      operationId: getOperatorPerformanceBudgetForecastHistory
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Reports outcomes for incidents initially opened by a projected budget exceedance and produces conservative read-only policy recommendations. A hit means the same incident later escalated to an actual exceedance; a cleared prediction may reflect successful mitigation and is not labelled a false positive. Observing predictions are excluded from hit-rate calculations. Recommendations require at least five decided outcomes and must pass impact preview before any manual update. Regressed effects can be acknowledged and closed through separately audited, admin-only endpoints that never update the budget.
      parameters:
        - { name: range, in: query, required: false, schema: { type: string, enum: [30d, 90d, 365d], default: 90d } }
        - { name: budget_id, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: limit, in: query, required: false, description: Limits returned detail items only; aggregate values cover the complete selected range., schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        "200":
          description: Read-only forecast outcome aggregates and recent details.
          content: { application/json: { schema: { $ref: "#/components/schemas/PerformanceBudgetForecastHistory" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/performance/budgets/report:
    get:
      operationId: exportOperatorPerformanceBudgetReview
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Downloads a privacy-safe periodic review containing current budget status, complete forecast outcomes, human decisions, effect observations, rollback plans, and post-rollback verification. It never evaluates budgets or sends notifications.
      parameters:
        - { name: range, in: query, required: false, schema: { type: string, enum: [30d, 90d, 365d], default: 90d } }
        - { name: format, in: query, required: false, schema: { type: string, enum: [json, markdown], default: json } }
      responses:
        "200":
          description: Downloadable JSON or Markdown budget-review report.
          headers:
            Content-Disposition: { schema: { type: string } }
          content:
            application/json: { schema: { $ref: "#/components/schemas/PerformanceBudgetReviewReport" } }
            text/markdown: { schema: { type: string } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/performance/budgets/evaluate:
    post:
      operationId: evaluateOperatorPerformanceBudgets
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Immediately evaluates every persisted budget and reconciles forecast-warning, actual-warning, critical, escalation, and recovery incident transitions. Each budget controls whether forecast incidents are enabled, the 1-90 day history window, and the 5-10000 completed-run sample minimum. The API also runs this evaluation periodically.
      responses:
        "200":
          description: Budget evaluation and incident-transition counts.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [evaluation]
                properties:
                  evaluation: { $ref: "#/components/schemas/PerformanceBudgetEvaluation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/ops/performance/budgets/{budget_id}:
    patch:
      operationId: updateOperatorPerformanceBudget
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Replaces mutable budget fields using an optimistic revision check and records the reason in the audit log. Requires the admin role.
      parameters:
        - { name: budget_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PerformanceBudgetMutation" } } }
      responses:
        "200":
          description: Updated budget.
          content:
            application/json:
              schema:
                type: object
                required: [budget]
                properties: { budget: { $ref: "#/components/schemas/PerformanceBudget" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    operatorBearerAuth:
      type: http
      scheme: bearer
      description: Uses either legacy OPERATOR_TOKEN or an operator account token. When operator MFA is enabled, clients must also send X-Operator-TOTP.
  responses:
    Unauthorized:
      description: Authentication is missing, invalid, expired, or revoked.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Forbidden:
      description: Authenticated operator role lacks permission for the requested action.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    ValidationError:
      description: Request failed domain validation.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    VersionPerformancePage:
      type: object
      additionalProperties: false
      required: [dimension, module, from, to, items]
      properties:
        dimension: { type: string, enum: [agent_version, model_profile] }
        module: { type: string }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        items: { type: array, items: { $ref: "#/components/schemas/VersionPerformance" } }
...
    ModelPerformancePage:
      type: object
      additionalProperties: false
      required: [module, from, to, items]
      properties:
        module: { type: string }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        items: { type: array, items: { $ref: "#/components/schemas/ModelPerformance" } }
...